requirements.txt
black==23.3.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
DateTime==5.1
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JSON==0.4.0
Flask-MySQLdb==1.0.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.3
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
tomli==2.0.1
Werkzeug==2.3.6
zope.interface==6.0
